pyproject.toml
[project]
name = "autoplayer"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "cel-ti", email = "96926560+cel-ti@users.noreply.github.com" }
]
dependencies = [
    "python-dotenv>=1.0.1",
    "zrcl>=0.3.1",
    "reldplayer[all]>=0.1.0",
    "pillow>=10.3.0",
    "discord-webhook>=1.3.1",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/autoplayer"]

[tool.ruff]
ignore = ["F401"]
